refactor: modularize lua block metadata
